Output fixed-duration noise regardless of input
Script now outputs a fixed-duration noise mask instead of basing noise duration on the duration of the input.

Previously, noise masks were set to the duration of the input recording + 400ms. Now, all noise masks generated are 1540ms in duration. (Reasoning: variable length of noise could provide information to the participant about the length of the word being masked.)